fix: bump compressed_morton_code to uint64

+++ cloudvolume/datasource/precomputed/image/common.py
@@ -68,15 +68,15 @@ def gridpoints(bbox, volume_bbox, chunk_size):
     yield Vec(x,y,z)
 
 def compressed_morton_code(pt, grid_size):
-  code = np.uint32(0)
+  code = np.uint64(0)
   num_bits = int(max([ np.ceil(np.log2(size)) for size in grid_size ]))
-  j = np.uint32(0)
-  one = np.uint32(1)
+  j = np.uint64(0)
+  one = np.uint64(1)
 
   for i in range(num_bits):
     for dim in range(3):
       if 2 ** i <= grid_size[dim]:
-        bit = (((np.uint32(pt[dim]) >> i) & one) << j)
+        bit = (((np.uint64(pt[dim]) >> np.uint64(i)) & one) << j)
         code |= bit
         j += one
